git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/SOC2007_WXWEB@47209 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/cocoa/pen.h
@@ -61,7 +61,7 @@ public:
     wxBitmap *GetStipple() const;
 
     WX_NSColor GetNSColor();
-    int GetCocoaLineDash(const float **pattern);
+    int GetCocoaLineDash(const CGFloat **pattern);
 };
 
 #endif // __WX_COCOA_PEN_H__
